pyproject.toml
[tool.poetry]
name = "test-assignment"
version = "0.1.0"
description = ""
authors = ["mikheyev9 <mikheyev9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.32"
aiogram = "^3.11.0"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.1"
colorlog = "^6.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
